change all cv number making 1 cv defalut working on designpage preview  remove signup btn after login
