+++ src/Components/Template/Cv2.jsx
@@ -1,94 +1,123 @@
-import React, { useState, useEffect, useRef } from "react";
+import React, { useState, useRef } from "react";
 import {
   ThemeProvider,
   createTheme,
   CssBaseline,
   Box,
   Typography,
+  IconButton,
+  Avatar,
   Grid,
-  List,
-  ListItem,
+  Chip,
+  Card,
+  CardContent,
   Button,
   Container,
-  IconButton,
-  CircularProgress,
-  Alert
+  List,
+  ListItem,
+  Tooltip
 } from "@mui/material";
-import { Print, PictureAsPdf, Edit, Delete } from "@mui/icons-material";
+import {
+  LinkedIn,
+  GitHub,
+  Twitter,
+  Download as DownloadIcon,
+  LightMode,
+  DarkMode,
+  Star,
+  Print,
+  Edit,
+  Delete,
+  Add
+} from "@mui/icons-material";
 import styled from "@emotion/styled";
-import { useReactToPrint } from "react-to-print";
 import html2canvas from "html2canvas";
-import jsPDF from "jspdf";
-import { apiUrl } from "../../utils/common";
+import { jsPDF } from "jspdf";
+import { useReactToPrint } from "react-to-print";
 
-const themeData = {
-  name: "Professional Light",
-  bg: "#FFFFFF",
-  text: "#2C3E50",
-  accent: "#2980B9",
-  font: "'Roboto Slab', serif",
-  body: "'Roboto', sans-serif",
-};
+// Themes configuration
+const themes = [
+  {
+    name: "Dark Charcoal",
+    bg: "#2C2C2C",
+    text: "#F5F5DC",
+    accent: "#3B82F6",
+    font: "Manrope, sans-serif",
+  },
+  {
+    name: "Light Cream",
+    bg: "#F5F5DC",
+    text: "#2C2C2C",
+    accent: "#3B82F6",
+    font: "Inter, sans-serif",
+  },
+  {
+    name: "Blue Gray",
+    bg: "#1E293B",
+    text: "#F5F5DC",
+    accent: "#3B82F6",
+    font: "Fira Sans, sans-serif",
+  },
+  {
+    name: "Soft Olive",
+    bg: "#556B2F",
+    text: "#F5F5DC",
+    accent: "#3B82F6",
+    font: "Source Sans Pro, sans-serif",
+  },
+];
 
+// Styled components
 const CVContainer = styled(Box)`
   width: 210mm;
   min-height: 297mm;
-  margin: 0 auto;
-  padding: 15mm;
-  position: relative;
-  box-shadow: 0 4px 20px rgba(0,0,0,0.08);
-  background-color: ${themeData.bg};
+  margin: auto;
+  padding: 20mm;
+  background: ${({ theme }) => theme.palette.background.default};
+  border-radius: 8px;
   box-sizing: border-box;
-  overflow: hidden;
+  position: relative;
 
   @media print {
     box-shadow: none;
-    padding: 0 !important;
-    margin: 0 !important;
+    padding: 20mm !important;
     width: 210mm !important;
     min-height: 297mm !important;
-    overflow: visible !important;
+    margin: 0 !important;
+    page-break-after: always;
+    background: white !important;
   }
 `;
 
-const Initial = styled(Typography)`
-  font-size: 12rem;
-  font-weight: bold;
-  opacity: 0.05;
-  position: absolute;
-  top: -30px;
-  right: 0;
-  z-index: 0;
-  line-height: 1;
-  pointer-events: none;
-
-  @media print {
-    opacity: 0.07;
-    font-size: 10rem;
-    top: -20px;
-  }
-`;
-
-const Section = styled(Box)`
-  margin-bottom: 1.5rem;
+const SectionCard = styled(Card)`
+  background: ${({ theme }) =>
+    theme.palette.mode === 'light'
+      ? 'rgba(255, 255, 255, 0.8)'
+      : 'rgba(255, 255, 255, 0.05)'};
+  border-radius: 12px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+  margin-bottom: 16px;
+  page-break-inside: avoid;
   position: relative;
-  z-index: 1;
+
+  &:hover .section-controls {
+    opacity: 1;
+  }
 
   @media print {
-    margin-bottom: 1rem;
-    page-break-inside: avoid;
+    box-shadow: none !important;
+    background: white !important;
+    color: black !important;
+    border: 1px solid #eee;
   }
 `;
 
 const SectionTitle = styled(Typography)`
   font-weight: 700 !important;
-  letter-spacing: 1px;
-  margin-bottom: 0.75rem !important;
+  margin-bottom: 1rem !important;
   position: relative;
   display: inline-block;
   padding-bottom: 4px;
-  text-transform: uppercase;
-  font-size: 1.1rem !important;
 
   &::after {
     content: "";
@@ -97,13 +126,49 @@ const SectionTitle = styled(Typography)`
     left: 0;
     width: 50px;
     height: 2px;
-    background: ${themeData.accent};
+    background: ${({ theme }) => theme.palette.primary.main};
+  }
+
+  @media print {
+    color: #3B82F6 !important;
+    &::after {
+      background: #3B82F6 !important;
+    }
   }
 `;
 
 const PrintHide = styled(Box)`
   @media print {
     display: none !important;
+  }
+`;
+
+const ExperienceItem = styled(Box)`
+  margin-bottom: 1.5rem;
+  position: relative;
+  padding-left: 16px;
+  border-left: 3px solid ${({ theme }) => theme.palette.primary.main};
+  
+  &::before {
+    content: "";
+    position: absolute;
+    left: -6px;
+    top: 6px;
+    width: 10px;
+    height: 10px;
+    border-radius: 50%;
+    background: ${({ theme }) => theme.palette.primary.main};
+  }
+
+  &:hover .edit-controls {
+    opacity: 1;
+  }
+
+  @media print {
+    border-left: 3px solid #3B82F6 !important;
+    &::before {
+      background: #3B82F6 !important;
+    }
   }
 `;
 
@@ -118,7 +183,7 @@ const EditControls = styled(Box)`
   position: absolute;
   top: -12px;
   right: -12px;
-  background: ${themeData.bg};
+  background: white;
   border-radius: 4px;
   box-shadow: 0 2px 8px rgba(0,0,0,0.15);
   opacity: 0;
@@ -131,10 +196,20 @@ const EditControls = styled(Box)`
   }
 `;
 
-const EditableText = styled(Box)`
-  position: relative;
-  &:hover .edit-controls {
-    opacity: 1;
+const SectionControls = styled(Box)`
+  position: absolute;
+  top: 8px;
+  right: 8px;
+  background: white;
+  border-radius: 4px;
+  box-shadow: 0 2px 8px rgba(0,0,0,0.15);
+  opacity: 0;
+  transition: opacity 0.2s ease;
+  z-index: 10;
+  display: flex;
+  
+  @media print {
+    display: none !important;
   }
 `;
 
@@ -147,237 +222,123 @@ const EditableImage = styled(Box)`
   }
 `;
 
-export default function Cv2() {
+// const EditableText = styled(Box)`
+//   position: relative;
+//   &:hover .edit-controls {
+//     opacity: 1;
+//   }
+// `;
+
+ export default function Cv2() {
+  const [themeIndex, setThemeIndex] = useState(1); // Default to light theme
+  const active = themes[themeIndex];
   const cvRef = useRef();
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
+  const [isGeneratingPDF, setIsGeneratingPDF] = useState(false);
   const [cvData, setCvData] = useState({
     personal: {
-      name: "",
-      title: "",
-      initial: ""
+      name: "Priya Desai",
+      title: "Senior Fullstack Developer",
+      summary: "Senior Fullstack Developer with 5+ years of experience building scalable web applications. Specialized in React ecosystems with expertise in design systems and component libraries.",
+      profileImage: "https://via.placeholder.com/150"
     },
     contact: {
-      items: []
+      phone: "+91 9876012345",
+      email: "priya.desai@email.com",
+      location: "Pune, Maharashtra",
+      social: [
+        { id: 1, icon: <LinkedIn />, url: "#" },
+        { id: 2, icon: <GitHub />, url: "#" },
+        { id: 3, icon: <Twitter />, url: "#" },
+      ]
     },
-    skills: {
-      items: []
-    },
-    languages: {
-      items: []
-    },
-    awards: {
-      items: []
-    },
-    achievements: {
-      items: []
-    },
-    interests: {
-      items: []
-    },
-    profile: {
-      content: ""
-    },
-    experience: [],
-    education: [],
-    certifications: {
-      items: []
-    },
-    projects: [],
-    profileImage: null
-  });
-
-  // Fetch data from API
-  useEffect(() => {
-    const fetchCVData = async () => {
-      try {
-        setLoading(true);
-        const response = await fetch(`${apiUrl}/defaultCv/mukesh_277`);
-
-        if (!response.ok) {
-          throw new Error(`HTTP error! status: ${response.status}`);
-        }
-
-        const data = await response.json();
-
-        // Log the API response to console
-        console.log("API Response:", data);
-
-        // Check if the API response has the expected structure
-        if (data.status === 'success' && data.fetchedCvInfo) {
-          // Transform API data to match our component structure
-          const transformedData = transformAPIData(data.fetchedCvInfo);
-          setCvData(transformedData);
-        } else {
-          throw new Error("Invalid API response structure");
-        }
-
-        setLoading(false);
-      } catch (err) {
-        // console.error("Error fetching CV data:", err);
-        setError(err.message);
-        setLoading(false);
+    skills: [
+      "React", "Next.js", "Node.js", "TypeScript",
+      "GraphQL", "Material UI", "Tailwind CSS", "Redux",
+      "Express", "MongoDB", "PostgreSQL", "AWS",
+      "Docker", "Jest", "Cypress", "CI/CD"
+    ],
+    experience: [
+      {
+        id: 1,
+        title: "Senior Fullstack Developer",
+        company: "ZYX Digital | 2022–Present",
+        description: "• Led development of enterprise SaaS platform serving 100K+ users\n• Created design system used across 15+ products\n• Reduced bundle size by 40% through code optimization"
+      },
+      {
+        id: 2,
+        title: "Frontend Developer",
+        company: "LMN Studio | 2020–2022",
+        description: "• Developed responsive UIs for e-commerce platforms\n• Implemented component library reducing dev time by 30%\n• Optimized performance achieving 95+ Lighthouse scores"
       }
-    };
-
-    fetchCVData();
-  }, []);
-
-  // Function to transform API data to our component structure
-  const transformAPIData = (apiData) => {
-    // Create initials from first and last name
-    const firstName = apiData.firstName || '';
-    const lastName = apiData.lastName || '';
-    const initials = `${firstName.charAt(0)}${lastName.charAt(0)}`;
-
-    // Create address string
-    const address = apiData.defaultCvInfo?.address
-      ? `${apiData.defaultCvInfo.address.city || ''}, ${apiData.defaultCvInfo.address.state || ''}, ${apiData.defaultCvInfo.address.country || ''}`
-      : '';
-
-    // Format date of birth
-    const dob = apiData.dob ? new Date(apiData.dob).toLocaleDateString() : '';
-
-    // Create contact items
-    const contactItems = [
-      { id: 1, content: apiData.phoneNo || '' },
-      { id: 2, content: apiData.email || '' },
-      { id: 3, content: address },
-      { id: 4, content: `DOB: ${dob}` },
-      { id: 5, content: `Gender: ${apiData.gender || ''}` },
-      ...(apiData.socialLinks?.map((link, index) => ({ id: 6 + index, content: link })) || [])
-    ].filter(item => item.content && item.content !== 'DOB: ' && item.content !== 'Gender: '); // Remove empty items
-
-    // Extract sections from API data
-    const sections = {};
-    if (apiData.sections && Array.isArray(apiData.sections)) {
-      apiData.sections.forEach(section => {
-        sections[section.name] = section.data;
-      });
-    }
-
-    return {
-      personal: {
-        name: `${firstName} ${lastName}`.trim(),
-        title: apiData.designation || '',
-        initial: initials
+    ],
+    education: [
+      {
+        id: 1,
+        degree: "Master of Computer Applications",
+        institution: "Pune University | 2017–2020",
+        details: "CGPA: 9.1/10"
       },
-      contact: {
-        items: contactItems
+      {
+        id: 2,
+        degree: "Bachelor of Science (Computer Science)",
+        institution: "Mumbai University | 2014–2017"
+      }
+    ],
+    projects: [
+      {
+        id: 1,
+        title: "Design System Library",
+        description: "Comprehensive React component library with 50+ components used across company products."
       },
-      skills: {
-        items: (sections.Skill || []).map((skill, index) => ({
-          id: index + 1,
-          content: `${skill.skill}${skill.rating ? ` (${skill.rating}/10)` : ''}`
-        }))
-      },
-      languages: {
-        items: (sections.Language || []).map((lang, index) => ({
-          id: index + 1,
-          content: `${lang.language} (${lang.proficiency})`
-        }))
-      },
-      awards: {
-        items: (sections.Award || []).map((award, index) => ({
-          id: index + 1,
-          title: award.title || '',
-          subtitle: `${award.issuer || ''}${award.date ? `, ${new Date(award.date).getFullYear()}` : ''}`
-        }))
-      },
-      achievements: {
-        items: (sections.Achievement || []).map((achievement, index) => ({
-          id: index + 1,
-          content: achievement
-        }))
-      },
-      interests: {
-        items: (sections.Interest || []).map((interest, index) => ({
-          id: index + 1,
-          content: interest
-        }))
-      },
-      profile: {
-        content: sections.Summary || ''
-      },
-      experience: (sections.Experience || []).map((exp, index) => ({
-        id: index + 1,
-        title: exp.jobTitle || '',
-        company: `${exp.company || ''}${exp.location ? `, ${exp.location}` : ''} | ${formatDateRange(exp.startDate, exp.endDate)}`,
-        description: exp.description || ''
-      })),
-      education: (sections.Education || []).map((edu, index) => ({
-        id: index + 1,
-        degree: `${edu.course || ''}${edu.fieldOfStudy ? ` in ${edu.fieldOfStudy}` : ''}`,
-        institution: `${edu.college || ''}${edu.location ? `, ${edu.location}` : ''} | ${formatDateRange(edu.startDate, edu.endDate)}${edu.grade ? `, Grade: ${edu.grade}` : ''}`
-      })),
-      certifications: {
-        items: (sections.Certification || []).map((cert, index) => ({
-          id: index + 1,
-          content: `${cert.name || ''}${cert.institute ? `, ${cert.institute}` : ''}${cert.issueDate ? `, ${new Date(cert.issueDate).getFullYear()}` : ''}`
-        }))
-      },
-      projects: (sections.Project || []).map((proj, index) => ({
-        id: index + 1,
-        title: proj.name || '',
-        description: proj.description || ''
-      })),
-      profileImage: apiData.profilePhoto || null
-    };
-  };
-
-  // Helper function to format date range
-  const formatDateRange = (startDate, endDate) => {
-    const formatDate = (dateString) => {
-      if (!dateString) return '';
-      const date = new Date(dateString);
-      return date.getFullYear();
-    };
-
-    const start = formatDate(startDate);
-    const end = endDate ? formatDate(endDate) : 'Present';
-
-    return start && end ? `${start}-${end}` : start || end;
-  };
+      {
+        id: 2,
+        title: "SaaS Analytics Dashboard",
+        description: "Real-time analytics platform processing 1M+ events daily."
+      }
+    ],
+    certifications: [
+      "Meta Frontend Professional Certificate (Coursera)",
+      "AWS Certified Developer - Associate",
+      "Google Cloud Professional Developer",
+      "React Advanced Concepts (Frontend Masters)"
+    ],
+    achievements: [
+      "Speaker at React Conf India 2023",
+      "Published 15+ technical articles on Medium"
+    ]
+  });
 
   const theme = createTheme({
     palette: {
-      mode: "light",
-      background: { default: themeData.bg, paper: themeData.bg },
-      text: { primary: themeData.text },
-      primary: { main: themeData.accent },
+      mode: themeIndex === 1 ? "light" : "dark",
+      background: { default: active.bg, paper: active.bg },
+      text: { primary: active.text },
+      primary: { main: active.accent },
     },
     typography: {
-      fontFamily: themeData.body,
-      h4: {
-        fontFamily: themeData.font,
-        fontWeight: 700,
-        letterSpacing: 1.5,
-        marginBottom: '0.5rem'
-      },
-      h5: {
-        fontFamily: themeData.font,
-        fontWeight: 600,
-        letterSpacing: 1
-      },
-      h6: {
-        fontFamily: themeData.font,
-        letterSpacing: 1.2,
-        fontWeight: 500
-      },
-      body1: {
-        lineHeight: 1.6,
-        fontSize: '0.95rem'
-      },
-      body2: {
-        lineHeight: 1.5,
-        fontSize: '0.9rem'
-      }
+      fontFamily: active.font,
+      h4: { fontWeight: 700, letterSpacing: 0.5 },
+      h5: { fontWeight: 600, letterSpacing: 0.5 },
+      body1: { lineHeight: 1.6 }
     },
     components: {
-      MuiDivider: {
+      MuiChip: {
         styleOverrides: {
           root: {
-            borderColor: themeData.accent + "40"
+            marginRight: 4,
+            marginBottom: 4,
+            '@media print': {
+              borderColor: '#3B82F6 !important',
+              color: 'black !important'
+            }
+          }
+        }
+      },
+      MuiLinearProgress: {
+        styleOverrides: {
+          root: {
+            height: 6,
+            borderRadius: 5
           }
         }
       }
@@ -385,30 +346,43 @@ export default function Cv2() {
   });
 
   // Handle text editing
-  const handleTextChange = (path, id, value) => {
+  const handleTextChange = (path, value) => {
     setCvData(prev => {
       const newData = { ...prev };
       let target = newData;
+      const segments = path.split('.');
 
-      path.split('.').forEach(segment => {
+      segments.slice(0, -1).forEach(segment => {
         target = target[segment];
       });
 
-      if (Array.isArray(target)) {
-        const index = target.findIndex(item => item.id === id);
-        if (index !== -1) {
-          target[index] = { ...target[index], content: value };
-        }
-      } else if (typeof target === 'object' && target !== null) {
-        target.content = value;
-      }
+      const lastSegment = segments[segments.length - 1];
+      target[lastSegment] = value;
 
       return newData;
     });
   };
 
-  // Handle complex object editing
-  const handleObjectFieldChange = (path, id, field, value) => {
+  // Handle array item editing
+  // const handleArrayItemChange = (path, index, value) => {
+  //   setCvData(prev => {
+  //     const newData = { ...prev };
+  //     let target = newData;
+  //     const segments = path.split('.');
+
+  //     segments.slice(0, -1).forEach(segment => {
+  //       target = target[segment];
+  //     });
+
+  //     const lastSegment = segments[segments.length - 1];
+  //     target[lastSegment][index] = value;
+
+  //     return newData;
+  //   });
+  // };
+
+  // Handle object item editing
+  const handleObjectItemChange = (path, id, field, value) => {
     setCvData(prev => {
       const newData = { ...prev };
       const targetArray = newData[path];
@@ -450,7 +424,10 @@ export default function Cv2() {
       reader.onloadend = () => {
         setCvData(prev => ({
           ...prev,
-          profileImage: reader.result
+          personal: {
+            ...prev.personal,
+            profileImage: reader.result
+          }
         }));
       };
       reader.readAsDataURL(file);
@@ -461,14 +438,107 @@ export default function Cv2() {
   const handleImageDelete = () => {
     setCvData(prev => ({
       ...prev,
-      profileImage: null
+      personal: {
+        ...prev.personal,
+        profileImage: null
+      }
     }));
   };
 
-  // Print functionality
+  // Add new item to a section
+  const addNewItem = (section) => {
+    const newItemId = Date.now();
+
+    setCvData(prev => {
+      if (section === 'experience') {
+        return {
+          ...prev,
+          experience: [
+            ...prev.experience,
+            {
+              id: newItemId,
+              title: "New Position",
+              company: "Company Name | Year",
+              description: "• Description point 1\n• Description point 2"
+            }
+          ]
+        };
+      } else if (section === 'projects') {
+        return {
+          ...prev,
+          projects: [
+            ...prev.projects,
+            {
+              id: newItemId,
+              title: "New Project",
+              description: "Project description"
+            }
+          ]
+        };
+      }
+      return prev;
+    });
+  };
+
+  // PDF Download functionality
+  const handleDownloadPDF = async () => {
+    setIsGeneratingPDF(true);
+    const originalTheme = themeIndex;
+
+    // Force light theme for PDF generation
+    if (originalTheme !== 1) {
+      setThemeIndex(1);
+      await new Promise(resolve => setTimeout(resolve, 500));
+    }
+
+    try {
+      const element = cvRef.current;
+      const canvas = await html2canvas(element, {
+        scale: 2,
+        logging: false,
+        useCORS: true,
+        backgroundColor: "#FFFFFF",
+        scrollX: 0,
+        scrollY: 0,
+        windowWidth: element.scrollWidth,
+        windowHeight: element.scrollHeight
+      });
+
+      const imgData = canvas.toDataURL('image/png');
+      const pdf = new jsPDF('p', 'mm', 'a4');
+      const imgWidth = 210;
+      const pageHeight = 297;
+      const imgHeight = (canvas.height * imgWidth) / canvas.width;
+
+      let heightLeft = imgHeight;
+      let position = 0;
+
+      pdf.addImage(imgData, 'PNG', 0, position, imgWidth, imgHeight);
+      heightLeft -= pageHeight;
+
+      while (heightLeft >= 0) {
+        position = heightLeft - imgHeight;
+        pdf.addPage();
+        pdf.addImage(imgData, 'PNG', 0, position, imgWidth, imgHeight);
+        heightLeft -= pageHeight;
+      }
+
+      pdf.save('professional-cv.pdf');
+    } catch (error) {
+      console.error('Error generating PDF:', error);
+    } finally {
+      if (originalTheme !== 1) {
+        setThemeIndex(originalTheme);
+      }
+      setIsGeneratingPDF(false);
+    }
+  };
+
+  const nextTheme = () =>
+    setThemeIndex((prev) => (prev + 1) % themes.length);
+
   const handlePrint = useReactToPrint({
     content: () => cvRef.current,
-    documentTitle: "Professional_CV",
     pageStyle: `
       @page {
         size: A4;
@@ -479,430 +549,481 @@ export default function Cv2() {
           width: 210mm;
           height: 297mm;
         }
-        body { 
-          -webkit-print-color-adjust: exact;
-          print-color-adjust: exact;
-          background: ${themeData.bg} !important;
+        body {
           margin: 0;
           padding: 0;
+          background: white !important;
+          -webkit-print-color-adjust: exact !important;
+          color-adjust: exact !important;
+          print-color-adjust: exact !important;
         }
-        ${CVContainer} {
-          box-shadow: none;
-          padding: 0 !important;
-          margin: 0 !important;
-          width: 210mm !important;
-          min-height: 297mm !important;
+        * {
+          -webkit-print-color-adjust: exact !important;
+          color-adjust: exact !important;
+          print-color-adjust: exact !important;
+        }
+        .MuiCard-root {
+          break-inside: avoid;
         }
       }
     `,
     onAfterPrint: () => console.log("Printed successfully!")
   });
 
-  // PDF Download functionality
-  const handleDownloadPDF = async () => {
-    const input = cvRef.current;
-    const canvas = await html2canvas(input, {
-      scale: 2,
-      logging: false,
-      useCORS: true,
-      scrollX: 0,
-      scrollY: 0,
-      windowWidth: input.scrollWidth,
-      windowHeight: input.scrollHeight
-    });
-
-    const imgData = canvas.toDataURL('image/png');
-    const pdf = new jsPDF({
-      orientation: 'portrait',
-      unit: 'mm',
-      format: 'a4'
-    });
-
-    const pdfWidth = pdf.internal.pageSize.getWidth();
-    const pdfHeight = pdf.internal.pageSize.getHeight();
-
-    // Calculate the scale to fit the content
-    const imgWidth = canvas.width;
-    const imgHeight = canvas.height;
-    const ratio = Math.min(pdfWidth / imgWidth, pdfHeight / imgHeight) * 0.95;
-    const imgX = (pdfWidth - imgWidth * ratio) / 2;
-    const imgY = 5;
-
-    pdf.addImage(imgData, 'PNG', imgX, imgY, imgWidth * ratio, imgHeight * ratio);
-    pdf.save('Professional_CV.pdf');
-  };
-
-  if (loading) {
-    return (
-      <Box display="flex" justifyContent="center" alignItems="center" minHeight="100vh">
-        <CircularProgress />
-      </Box>
-    );
-  }
-
-  if (error) {
-    return (
-      <Container maxWidth="md" sx={{ mt: 4 }}>
-        <Alert severity="error">
-          Error loading CV data: {error}
-        </Alert>
-      </Container>
-    );
-  }
-
   return (
     <ThemeProvider theme={theme}>
       <CssBaseline />
-      <Container
-        maxWidth={false}
-        sx={{
-          padding: '20px 0',
-          '@media print': {
-            padding: 0,
-            margin: 0,
-            width: '100%',
-            height: '100%'
-          },
-        }}
-      >
-        <PrintHide
-          display="flex"
-          justifyContent="center"
-          mb={3}
-          flexWrap="wrap"
-          gap={2}
-          sx={{
-            position: 'relative',
-            zIndex: 10,
-          }}
-        >
+      <Container sx={{ '@media print': { padding: '0 !important' } }}>
+        <PrintHide display="flex" justifyContent="flex-end" mb={2}>
+          <IconButton onClick={nextTheme} color="primary" sx={{ mr: 1 }}>
+            {themeIndex === 1 ? <DarkMode /> : <LightMode />}
+          </IconButton>
           <Button
-            variant="contained"
-            color="primary"
+            variant="outlined"
             startIcon={<Print />}
             onClick={handlePrint}
-            sx={{ minWidth: '180px' }}
+            sx={{ mr: 1 }}
           >
-            Print CV
+            Print
           </Button>
           <Button
             variant="contained"
             color="primary"
-            startIcon={<PictureAsPdf />}
+            startIcon={<DownloadIcon />}
             onClick={handleDownloadPDF}
-            sx={{ minWidth: '180px' }}
+            disabled={isGeneratingPDF}
           >
-            Download PDF
+            {isGeneratingPDF ? 'Generating...' : 'Download PDF'}
           </Button>
         </PrintHide>
 
-        <Box display="flex" justifyContent="center" sx={{ '@media print': { margin: 0 } }}>
-          <CVContainer ref={cvRef}>
-            <Initial>{cvData.personal.initial}</Initial>
+        <CVContainer ref={cvRef}>
+          {/* Header */}
+          <Box display="flex" justifyContent="space-between" alignItems="center" mb={3}>
+            <EditableWrapper>
+              <Typography variant="h4" color="primary" sx={{ '@media print': { color: '#3B82F6 !important' } }}>
+                {cvData.personal.name}
+              </Typography>
+              <EditControls className="edit-controls">
+                <IconButton size="small" onClick={() => {
+                  const newValue = prompt("Edit name", cvData.personal.name);
+                  if (newValue !== null) {
+                    handleTextChange('personal.name', newValue);
+                  }
+                }}>
+                  <Edit fontSize="small" />
+                </IconButton>
+              </EditControls>
+            </EditableWrapper>
+            <EditableWrapper>
+              <Typography variant="h6" color="textSecondary" sx={{ '@media print': { color: 'black !important' } }}>
+                {cvData.personal.title}
+              </Typography>
+              <EditControls className="edit-controls">
+                <IconButton size="small" onClick={() => {
+                  const newValue = prompt("Edit title", cvData.personal.title);
+                  if (newValue !== null) {
+                    handleTextChange('personal.title', newValue);
+                  }
+                }}>
+                  <Edit fontSize="small" />
+                </IconButton>
+              </EditControls>
+            </EditableWrapper>
+          </Box>
 
-            <Box mb={4} position="relative" zIndex={1}>
-              <EditableWrapper>
-                <Typography variant="h4" color="primary">
-                  {cvData.personal.name}
-                </Typography>
-                <EditControls className="edit-controls">
-                  <IconButton size="small" onClick={() => {
-                    const newName = prompt("Edit name", cvData.personal.name);
-                    if (newName) {
-                      setCvData(prev => ({
-                        ...prev,
-                        personal: {
-                          ...prev.personal,
-                          name: newName
-                        }
-                      }));
-                    }
-                  }}>
-                    <Edit fontSize="small" />
-                  </IconButton>
-                </EditControls>
-              </EditableWrapper>
+          <Grid container spacing={3}>
+            {/* Left Column */}
+            <Grid item xs={12} md={5}>
+              <SectionCard>
+                <CardContent>
+                  <SectionControls className="section-controls">
+                    <Tooltip title="Edit section">
+                      <IconButton size="small">
+                        <Edit fontSize="small" />
+                      </IconButton>
+                    </Tooltip>
+                  </SectionControls>
+                  <Grid container spacing={2} alignItems="center">
+                    <Grid item xs={12} md={5}>
+                      {cvData.personal.profileImage ? (
+                        <EditableImage>
+                          <Avatar
+                            src={cvData.personal.profileImage}
+                            sx={{
+                              width: 120,
+                              height: 120,
+                              border: `3px solid ${active.accent}`,
+                              '@media print': { borderColor: '#3B82F6 !important' }
+                            }}
+                          />
+                          <EditControls className="edit-controls">
+                            <IconButton size="small" onClick={() => document.getElementById('image-upload').click()}>
+                              <Edit fontSize="small" />
+                            </IconButton>
+                            <IconButton size="small" onClick={handleImageDelete}>
+                              <Delete fontSize="small" />
+                            </IconButton>
+                            <input
+                              id="image-upload"
+                              type="file"
+                              accept="image/*"
+                              style={{ display: 'none' }}
+                              onChange={handleImageUpload}
+                            />
+                          </EditControls>
+                        </EditableImage>
+                      ) : (
+                        <Button
+                          variant="outlined"
+                          onClick={() => document.getElementById('image-upload').click()}
+                          sx={{ width: 120, height: 120 }}
+                        >
+                          Add Photo
+                        </Button>
+                      )}
+                    </Grid>
+                    <Grid item xs={12} md={7}>
+                      <EditableWrapper>
+                        <Typography variant="body1" sx={{ '@media print': { color: 'black !important' } }}>
+                          <Box component="span" fontWeight="bold">Phone:</Box> {cvData.contact.phone}
+                        </Typography>
+                        <EditControls className="edit-controls">
+                          <IconButton size="small" onClick={() => {
+                            const newValue = prompt("Edit phone", cvData.contact.phone);
+                            if (newValue !== null) {
+                              handleTextChange('contact.phone', newValue);
+                            }
+                          }}>
+                            <Edit fontSize="small" />
+                          </IconButton>
+                        </EditControls>
+                      </EditableWrapper>
 
-              <EditableWrapper>
-                <Typography variant="h6" sx={{ letterSpacing: 1.5 }}>
-                  {cvData.personal.title}
-                </Typography>
-                <EditControls className="edit-controls">
-                  <IconButton size="small" onClick={() => {
-                    const newTitle = prompt("Edit title", cvData.personal.title);
-                    if (newTitle) {
-                      setCvData(prev => ({
-                        ...prev,
-                        personal: {
-                          ...prev.personal,
-                          title: newTitle
-                        }
-                      }));
-                    }
-                  }}>
-                    <Edit fontSize="small" />
-                  </IconButton>
-                </EditControls>
-              </EditableWrapper>
-            </Box>
+                      <EditableWrapper>
+                        <Typography variant="body1" sx={{ '@media print': { color: 'black !important' } }}>
+                          <Box component="span" fontWeight="bold">Email:</Box> {cvData.contact.email}
+                        </Typography>
+                        <EditControls className="edit-controls">
+                          <IconButton size="small" onClick={() => {
+                            const newValue = prompt("Edit email", cvData.contact.email);
+                            if (newValue !== null) {
+                              handleTextChange('contact.email', newValue);
+                            }
+                          }}>
+                            <Edit fontSize="small" />
+                          </IconButton>
+                        </EditControls>
+                      </EditableWrapper>
 
-            <Grid container spacing={4}>
-              <Grid item xs={12} md={4}>
-                <Section>
+                      <EditableWrapper>
+                        <Typography variant="body1" sx={{ '@media print': { color: 'black !important' } }}>
+                          <Box component="span" fontWeight="bold">Location:</Box> {cvData.contact.location}
+                        </Typography>
+                        <EditControls className="edit-controls">
+                          <IconButton size="small" onClick={() => {
+                            const newValue = prompt("Edit location", cvData.contact.location);
+                            if (newValue !== null) {
+                              handleTextChange('contact.location', newValue);
+                            }
+                          }}>
+                            <Edit fontSize="small" />
+                          </IconButton>
+                        </EditControls>
+                      </EditableWrapper>
+
+                      <Box mt={1} sx={{ '@media print': { display: 'none' } }}>
+                        {cvData.contact.social.map((item) => (
+                          <IconButton key={item.id} color="primary">
+                            {item.icon}
+                          </IconButton>
+                        ))}
+                      </Box>
+                    </Grid>
+                  </Grid>
+                </CardContent>
+              </SectionCard>
+
+              <SectionCard>
+                <CardContent>
                   <SectionTitle variant="h5" color="primary">
-                    Contact
+                    Skills
                   </SectionTitle>
-                  <Box>
-                    {cvData.profileImage && (
-                      <EditableImage>
-                        <img
-                          src={cvData.profileImage}
-                          alt="Profile"
-                          style={{
-                            width: 80,
-                            height: 80,
-                            borderRadius: '50%',
-                            objectFit: 'cover',
-                            marginBottom: 8
-                          }}
-                        />
-                        <EditControls className="edit-controls">
-                          <IconButton size="small" onClick={() => document.getElementById('image-upload').click()}>
-                            <Edit fontSize="small" />
-                          </IconButton>
-                          <IconButton size="small" onClick={handleImageDelete}>
-                            <Delete fontSize="small" />
-                          </IconButton>
-                        </EditControls>
-                        <input
-                          id="image-upload"
-                          type="file"
-                          accept="image/*"
-                          style={{ display: 'none' }}
-                          onChange={handleImageUpload}
-                        />
-                      </EditableImage>
-                    )}
-                    {cvData.contact.items.map((item) => (
-                      <EditableText key={item.id} sx={{ position: 'relative' }}>
-                        <Typography>{item.content}</Typography>
-                        <EditControls className="edit-controls">
-                          <IconButton size="small" onClick={() => {
-                            const newValue = prompt("Edit contact", item.content);
-                            if (newValue !== null) {
-                              handleTextChange('contact.items', item.id, newValue);
-                            }
-                          }}>
-                            <Edit fontSize="small" />
-                          </IconButton>
-                          <IconButton size="small" onClick={() => handleDeleteItem('contact.items', item.id)}>
-                            <Delete fontSize="small" />
-                          </IconButton>
-                        </EditControls>
-                      </EditableText>
-                    ))}
-                  </Box>
-                </Section>
-
-                {cvData.skills.items.length > 0 && (
-                  <Section>
-                    <SectionTitle variant="h5" color="primary">
-                      Skills
-                    </SectionTitle>
-                    <List dense sx={{ py: 0 }}>
-                      {cvData.skills.items.map((skill) => (
-                        <ListItem key={skill.id} sx={{ py: 0.25, px: 0, position: 'relative' }}>
-                          <EditableText>
-                            • {skill.content}
-                            <EditControls className="edit-controls">
-                              <IconButton size="small" onClick={() => {
-                                const newValue = prompt("Edit skill", skill.content);
-                                if (newValue !== null) {
-                                  handleTextChange('skills.items', skill.id, newValue);
-                                }
-                              }}>
-                                <Edit fontSize="small" />
-                              </IconButton>
-                              <IconButton size="small" onClick={() => handleDeleteItem('skills.items', skill.id)}>
-                                <Delete fontSize="small" />
-                              </IconButton>
-                            </EditControls>
-                          </EditableText>
-                        </ListItem>
-                      ))}
-                    </List>
-                  </Section>
-                )}
-
-                {cvData.languages.items.length > 0 && (
-                  <Section>
-                    <SectionTitle variant="h5" color="primary">
-                      Languages
-                    </SectionTitle>
-                    {cvData.languages.items.map((lang) => (
-                      <EditableText key={lang.id} sx={{ position: 'relative' }}>
-                        <Typography>{lang.content}</Typography>
-                        <EditControls className="edit-controls">
-                          <IconButton size="small" onClick={() => {
-                            const newValue = prompt("Edit language", lang.content);
-                            if (newValue !== null) {
-                              handleTextChange('languages.items', lang.id, newValue);
-                            }
-                          }}>
-                            <Edit fontSize="small" />
-                          </IconButton>
-                          <IconButton size="small" onClick={() => handleDeleteItem('languages.items', lang.id)}>
-                            <Delete fontSize="small" />
-                          </IconButton>
-                        </EditControls>
-                      </EditableText>
-                    ))}
-                  </Section>
-                )}
-
-                {cvData.awards.items.length > 0 && (
-                  <Section>
-                    <SectionTitle variant="h5" color="primary">
-                      Awards
-                    </SectionTitle>
-                    <Box>
-                      {cvData.awards.items.map((award) => (
-                        <Box key={award.id} sx={{ position: 'relative', mb: 1 }}>
-                          <EditableText>
-                            <Typography fontWeight={500}>{award.title}</Typography>
-                            <EditControls className="edit-controls">
-                              <IconButton size="small" onClick={() => {
-                                const newValue = prompt("Edit award title", award.title);
-                                if (newValue !== null) {
-                                  setCvData(prev => {
-                                    const newItems = prev.awards.items.map(item =>
-                                      item.id === award.id ? { ...item, title: newValue } : item
-                                    );
-                                    return {
-                                      ...prev,
-                                      awards: { ...prev.awards, items: newItems }
-                                    };
-                                  });
-                                }
-                              }}>
-                                <Edit fontSize="small" />
-                              </IconButton>
-                              <IconButton size="small" onClick={() => {
+                  <SectionControls className="section-controls">
+                    <Tooltip title="Add skill">
+                      <IconButton size="small" onClick={() => {
+                        const newSkill = prompt("Add new skill");
+                        if (newSkill) {
+                          setCvData(prev => ({
+                            ...prev,
+                            skills: [...prev.skills, newSkill]
+                          }));
+                        }
+                      }}>
+                        <Add fontSize="small" />
+                      </IconButton>
+                    </Tooltip>
+                  </SectionControls>
+                  <Grid container spacing={1}>
+                    {cvData.skills.map((skill, index) => (
+                      <Grid item xs={6} key={index}>
+                        <EditableWrapper>
+                          <Chip
+                            label={skill}
+                            color="primary"
+                            variant="outlined"
+                            sx={{ '@media print': { color: 'black !important', borderColor: '#3B82F6 !important' } }}
+                          />
+                          <EditControls className="edit-controls">
+                            <IconButton size="small" onClick={() => {
+                              const newValue = prompt("Edit skill", skill);
+                              if (newValue !== null) {
+                                const newSkills = [...cvData.skills];
+                                newSkills[index] = newValue;
                                 setCvData(prev => ({
                                   ...prev,
-                                  awards: {
-                                    ...prev.awards,
-                                    items: prev.awards.items.filter(item => item.id !== award.id)
-                                  }
+                                  skills: newSkills
                                 }));
-                              }}>
-                                <Delete fontSize="small" />
-                              </IconButton>
-                            </EditControls>
-                          </EditableText>
-                          <EditableText>
-                            <Typography variant="body2">{award.subtitle}</Typography>
-                            <EditControls className="edit-controls">
-                              <IconButton size="small" onClick={() => {
-                                const newValue = prompt("Edit award subtitle", award.subtitle);
-                                if (newValue !== null) {
-                                  setCvData(prev => {
-                                    const newItems = prev.awards.items.map(item =>
-                                      item.id === award.id ? { ...item, subtitle: newValue } : item
-                                    );
-                                    return {
-                                      ...prev,
-                                      awards: { ...prev.awards, items: newItems }
-                                    };
-                                  });
-                                }
-                              }}>
-                                <Edit fontSize="small" />
-                              </IconButton>
-                            </EditControls>
-                          </EditableText>
-                        </Box>
-                      ))}
-                    </Box>
-                  </Section>
-                )}
+                              }
+                            }}>
+                              <Edit fontSize="small" />
+                            </IconButton>
+                            <IconButton size="small" onClick={() => {
+                              setCvData(prev => ({
+                                ...prev,
+                                skills: prev.skills.filter((_, i) => i !== index)
+                              }));
+                            }}>
+                              <Delete fontSize="small" />
+                            </IconButton>
+                          </EditControls>
+                        </EditableWrapper>
+                      </Grid>
+                    ))}
+                  </Grid>
+                </CardContent>
+              </SectionCard>
 
-                {cvData.achievements.items.length > 0 && (
-                  <Section>
-                    <SectionTitle variant="h5" color="primary">
-                      Achievements
-                    </SectionTitle>
-                    <List dense sx={{ py: 0 }}>
-                      {cvData.achievements.items.map((achievement) => (
-                        <ListItem key={achievement.id} sx={{ py: 0.25, px: 0, position: 'relative' }}>
-                          <EditableText>
-                            • {achievement.content}
-                            <EditControls className="edit-controls">
-                              <IconButton size="small" onClick={() => {
-                                const newValue = prompt("Edit achievement", achievement.content);
-                                if (newValue !== null) {
-                                  handleTextChange('achievements.items', achievement.id, newValue);
-                                }
-                              }}>
-                                <Edit fontSize="small" />
-                              </IconButton>
-                              <IconButton size="small" onClick={() => handleDeleteItem('achievements.items', achievement.id)}>
-                                <Delete fontSize="small" />
-                              </IconButton>
-                            </EditControls>
-                          </EditableText>
-                        </ListItem>
-                      ))}
-                    </List>
-                  </Section>
-                )}
+              <SectionCard>
+                <CardContent>
+                  <SectionTitle variant="h5" color="primary">
+                    Certifications
+                  </SectionTitle>
+                  <SectionControls className="section-controls">
+                    <Tooltip title="Add certification">
+                      <IconButton size="small" onClick={() => {
+                        const newCert = prompt("Add new certification");
+                        if (newCert) {
+                          setCvData(prev => ({
+                            ...prev,
+                            certifications: [...prev.certifications, newCert]
+                          }));
+                        }
+                      }}>
+                        <Add fontSize="small" />
+                      </IconButton>
+                    </Tooltip>
+                  </SectionControls>
+                  <List dense>
+                    {cvData.certifications.map((cert, index) => (
+                      <ListItem key={index} sx={{ py: 0.5, position: 'relative' }}>
+                        <EditableWrapper>
+                          <Typography>• {cert}</Typography>
+                          <EditControls className="edit-controls">
+                            <IconButton size="small" onClick={() => {
+                              const newValue = prompt("Edit certification", cert);
+                              if (newValue !== null) {
+                                const newCerts = [...cvData.certifications];
+                                newCerts[index] = newValue;
+                                setCvData(prev => ({
+                                  ...prev,
+                                  certifications: newCerts
+                                }));
+                              }
+                            }}>
+                              <Edit fontSize="small" />
+                            </IconButton>
+                            <IconButton size="small" onClick={() => {
+                              setCvData(prev => ({
+                                ...prev,
+                                certifications: prev.certifications.filter((_, i) => i !== index)
+                              }));
+                            }}>
+                              <Delete fontSize="small" />
+                            </IconButton>
+                          </EditControls>
+                        </EditableWrapper>
+                      </ListItem>
+                    ))}
+                  </List>
+                </CardContent>
+              </SectionCard>
 
-                {cvData.interests.items.length > 0 && (
-                  <Section>
-                    <SectionTitle variant="h5" color="primary">
-                      Interests
-                    </SectionTitle>
-                    <List dense sx={{ py: 0 }}>
-                      {cvData.interests.items.map((interest) => (
-                        <ListItem key={interest.id} sx={{ py: 0.25, px: 0, position: 'relative' }}>
-                          <EditableText>
-                            • {interest.content}
-                            <EditControls className="edit-controls">
-                              <IconButton size="small" onClick={() => {
-                                const newValue = prompt("Edit interest", interest.content);
-                                if (newValue !== null) {
-                                  handleTextChange('interests.items', interest.id, newValue);
-                                }
-                              }}>
-                                <Edit fontSize="small" />
-                              </IconButton>
-                              <IconButton size="small" onClick={() => handleDeleteItem('interests.items', interest.id)}>
-                                <Delete fontSize="small" />
-                              </IconButton>
-                            </EditControls>
-                          </EditableText>
-                        </ListItem>
-                      ))}
-                    </List>
-                  </Section>
-                )}
-              </Grid>
+              <SectionCard>
+                <CardContent>
+                  <SectionTitle variant="h5" color="primary">
+                    Achievements
+                  </SectionTitle>
+                  <SectionControls className="section-controls">
+                    <Tooltip title="Add achievement">
+                      <IconButton size="small" onClick={() => {
+                        const newAchievement = prompt("Add new achievement");
+                        if (newAchievement) {
+                          setCvData(prev => ({
+                            ...prev,
+                            achievements: [...prev.achievements, newAchievement]
+                          }));
+                        }
+                      }}>
+                        <Add fontSize="small" />
+                      </IconButton>
+                    </Tooltip>
+                  </SectionControls>
+                  <Box mt={1}>
+                    {cvData.achievements.map((achievement, index) => (
+                      <Box key={index} display="flex" alignItems="center" mb={1} position="relative">
+                        <EditableWrapper>
+                          <Star color="primary" sx={{ mr: 1 }} />
+                          <Typography>{achievement}</Typography>
+                          <EditControls className="edit-controls">
+                            <IconButton size="small" onClick={() => {
+                              const newValue = prompt("Edit achievement", achievement);
+                              if (newValue !== null) {
+                                const newAchievements = [...cvData.achievements];
+                                newAchievements[index] = newValue;
+                                setCvData(prev => ({
+                                  ...prev,
+                                  achievements: newAchievements
+                                }));
+                              }
+                            }}>
+                              <Edit fontSize="small" />
+                            </IconButton>
+                            <IconButton size="small" onClick={() => {
+                              setCvData(prev => ({
+                                ...prev,
+                                achievements: prev.achievements.filter((_, i) => i !== index)
+                              }));
+                            }}>
+                              <Delete fontSize="small" />
+                            </IconButton>
+                          </EditControls>
+                        </EditableWrapper>
+                      </Box>
+                    ))}
+                  </Box>
+                </CardContent>
+              </SectionCard>
+            </Grid>
 
-              <Grid item xs={12} md={8}>
-                {cvData.profile.content && (
-                  <Section>
+            {/* Right Column */}
+            <Grid item xs={12} md={7}>
+              <SectionCard>
+                <CardContent>
+                  <SectionTitle variant="h5" color="primary">
+                    Professional Summary
+                  </SectionTitle>
+                  <EditableWrapper>
+                    <Typography sx={{ '@media print': { color: 'black !important' } }}>
+                      {cvData.personal.summary}
+                    </Typography>
+                    <EditControls className="edit-controls">
+                      <IconButton size="small" onClick={() => {
+                        const newValue = prompt("Edit summary", cvData.personal.summary);
+                        if (newValue !== null) {
+                          handleTextChange('personal.summary', newValue);
+                        }
+                      }}>
+                        <Edit fontSize="small" />
+                      </IconButton>
+                    </EditControls>
+                  </EditableWrapper>
+                </CardContent>
+              </SectionCard>
+
+              <SectionCard>
+                <CardContent>
+                  <Box display="flex" justifyContent="space-between" alignItems="center">
                     <SectionTitle variant="h5" color="primary">
-                      Profile
+                      Experience
                     </SectionTitle>
-                    <EditableText sx={{ position: 'relative' }}>
-                      <Typography>{cvData.profile.content}</Typography>
-                      <EditControls className="edit-controls">
+                    <Tooltip title="Add experience">
+                      <IconButton onClick={() => addNewItem('experience')}>
+                        <Add />
+                      </IconButton>
+                    </Tooltip>
+                  </Box>
+
+                  {cvData.experience.map((exp) => (
+                    <ExperienceItem key={exp.id}>
+                      <EditControls className="edit-controls" sx={{ top: -8, right: -8 }}>
                         <IconButton size="small" onClick={() => {
-                          const newValue = prompt("Edit profile", cvData.profile.content);
-                          if (newValue !== null) {
+                          const newTitle = prompt("Edit position", exp.title);
+                          if (newTitle !== null) {
+                            handleObjectItemChange('experience', exp.id, 'title', newTitle);
+                          }
+                        }}>
+                          <Edit fontSize="small" />
+                        </IconButton>
+                        <IconButton size="small" onClick={() => handleDeleteItem('experience', exp.id)}>
+                          <Delete fontSize="small" />
+                        </IconButton>
+                      </EditControls>
+
+                      <EditableWrapper>
+                        <Typography variant="subtitle1" fontWeight={600}>{exp.title}</Typography>
+                      </EditableWrapper>
+
+                      <EditableWrapper>
+                        <Typography color="primary" fontStyle="italic">{exp.company}</Typography>
+                        <EditControls className="edit-controls">
+                          <IconButton size="small" onClick={() => {
+                            const newValue = prompt("Edit company", exp.company);
+                            if (newValue !== null) {
+                              handleObjectItemChange('experience', exp.id, 'company', newValue);
+                            }
+                          }}>
+                            <Edit fontSize="small" />
+                          </IconButton>
+                        </EditControls>
+                      </EditableWrapper>
+
+                      <EditableWrapper sx={{ mt: 1 }}>
+                        <Typography variant="body2" whiteSpace="pre-line">
+                          {exp.description}
+                        </Typography>
+                        <EditControls className="edit-controls">
+                          <IconButton size="small" onClick={() => {
+                            const newValue = prompt("Edit description", exp.description);
+                            if (newValue !== null) {
+                              handleObjectItemChange('experience', exp.id, 'description', newValue);
+                            }
+                          }}>
+                            <Edit fontSize="small" />
+                          </IconButton>
+                        </EditControls>
+                      </EditableWrapper>
+                    </ExperienceItem>
+                  ))}
+                </CardContent>
+              </SectionCard>
+
+              <SectionCard>
+                <CardContent>
+                  <SectionTitle variant="h5" color="primary">
+                    Education
+                  </SectionTitle>
+                  {cvData.education.map((edu, index) => (
+                    <Box key={index} mb={2} position="relative">
+                      <EditControls className="edit-controls" sx={{ top: -8, right: -8 }}>
+                        <IconButton size="small" onClick={() => {
+                          const newDegree = prompt("Edit degree", edu.degree);
+                          if (newDegree !== null) {
+                            const newEducation = [...cvData.education];
+                            newEducation[index] = { ...newEducation[index], degree: newDegree };
                             setCvData(prev => ({
                               ...prev,
-                              profile: { content: newValue }
+                              education: newEducation
                             }));
                           }
                         }}>
@@ -911,204 +1032,115 @@ export default function Cv2() {
                         <IconButton size="small" onClick={() => {
                           setCvData(prev => ({
                             ...prev,
-                            profile: { content: "" }
+                            education: prev.education.filter((_, i) => i !== index)
                           }));
                         }}>
                           <Delete fontSize="small" />
                         </IconButton>
                       </EditControls>
-                    </EditableText>
-                  </Section>
-                )}
 
-                {cvData.experience.length > 0 && (
-                  <Section>
-                    <SectionTitle variant="h5" color="primary">
-                      Experience
-                    </SectionTitle>
-                    {cvData.experience.map((exp) => (
-                      <Box key={exp.id} mb={3} sx={{ position: 'relative' }}>
-                        <EditControls className="edit-controls" sx={{ top: -8, right: -8 }}>
+                      <EditableWrapper>
+                        <Typography fontWeight={600}>{edu.degree}</Typography>
+                      </EditableWrapper>
+
+                      <EditableWrapper>
+                        <Typography>{edu.institution}</Typography>
+                        <EditControls className="edit-controls">
                           <IconButton size="small" onClick={() => {
-                            const newTitle = prompt("Edit position", exp.title);
-                            if (newTitle !== null) {
-                              handleObjectFieldChange('experience', exp.id, 'title', newTitle);
+                            const newValue = prompt("Edit institution", edu.institution);
+                            if (newValue !== null) {
+                              const newEducation = [...cvData.education];
+                              newEducation[index] = { ...newEducation[index], institution: newValue };
+                              setCvData(prev => ({
+                                ...prev,
+                                education: newEducation
+                              }));
                             }
                           }}>
                             <Edit fontSize="small" />
                           </IconButton>
-                          <IconButton size="small" onClick={() => handleDeleteItem('experience', exp.id)}>
-                            <Delete fontSize="small" />
-                          </IconButton>
                         </EditControls>
+                      </EditableWrapper>
 
-                        <EditableText>
-                          <Typography fontWeight={600}>{exp.title}</Typography>
-                        </EditableText>
-
-                        <EditableText>
-                          <Typography color="primary" fontStyle="italic">{exp.company}</Typography>
+                      {edu.details && (
+                        <EditableWrapper>
+                          <Typography color="textSecondary">{edu.details}</Typography>
                           <EditControls className="edit-controls">
                             <IconButton size="small" onClick={() => {
-                              const newValue = prompt("Edit company", exp.company);
+                              const newValue = prompt("Edit details", edu.details);
                               if (newValue !== null) {
-                                handleObjectFieldChange('experience', exp.id, 'company', newValue);
+                                const newEducation = [...cvData.education];
+                                newEducation[index] = { ...newEducation[index], details: newValue };
+                                setCvData(prev => ({
+                                  ...prev,
+                                  education: newEducation
+                                }));
                               }
                             }}>
                               <Edit fontSize="small" />
                             </IconButton>
                           </EditControls>
-                        </EditableText>
+                        </EditableWrapper>
+                      )}
+                    </Box>
+                  ))}
+                </CardContent>
+              </SectionCard>
 
-                        <EditableText sx={{ mt: 1 }}>
-                          <Typography variant="body2" whiteSpace="pre-line">
-                            {exp.description}
-                          </Typography>
-                          <EditControls className="edit-controls">
-                            <IconButton size="small" onClick={() => {
-                              const newValue = prompt("Edit description", exp.description);
-                              if (newValue !== null) {
-                                handleObjectFieldChange('experience', exp.id, 'description', newValue);
-                              }
-                            }}>
-                              <Edit fontSize="small" />
-                            </IconButton>
-                          </EditControls>
-                        </EditableText>
-                      </Box>
-                    ))}
-                  </Section>
-                )}
-
-                <Grid container spacing={2}>
-                  {cvData.education.length > 0 && (
-                    <Grid item xs={12} sm={6}>
-                      <Section>
-                        <SectionTitle variant="h5" color="primary">
-                          Education
-                        </SectionTitle>
-                        <Box>
-                          {cvData.education.map((edu) => (
-                            <Box key={edu.id} sx={{ position: 'relative', mb: 1.5 }}>
-                              <EditControls className="edit-controls" sx={{ top: -8, right: -8 }}>
-                                <IconButton size="small" onClick={() => handleDeleteItem('education', edu.id)}>
-                                  <Delete fontSize="small" />
-                                </IconButton>
-                              </EditControls>
-
-                              <EditableText>
-                                <Typography fontWeight={600}>{edu.degree}</Typography>
-                                <EditControls className="edit-controls">
-                                  <IconButton size="small" onClick={() => {
-                                    const newValue = prompt("Edit degree", edu.degree);
-                                    if (newValue !== null) {
-                                      handleObjectFieldChange('education', edu.id, 'degree', newValue);
-                                    }
-                                  }}>
-                                    <Edit fontSize="small" />
-                                  </IconButton>
-                                </EditControls>
-                              </EditableText>
-
-                              <EditableText>
-                                <Typography variant="body2">{edu.institution}</Typography>
-                                <EditControls className="edit-controls">
-                                  <IconButton size="small" onClick={() => {
-                                    const newValue = prompt("Edit institution", edu.institution);
-                                    if (newValue !== null) {
-                                      handleObjectFieldChange('education', edu.id, 'institution', newValue);
-                                    }
-                                  }}>
-                                    <Edit fontSize="small" />
-                                  </IconButton>
-                                </EditControls>
-                              </EditableText>
-                            </Box>
-                          ))}
-                        </Box>
-                      </Section>
-                    </Grid>
-                  )}
-
-                  {cvData.certifications.items.length > 0 && (
-                    <Grid item xs={12} sm={6}>
-                      <Section>
-                        <SectionTitle variant="h5" color="primary">
-                          Certifications
-                        </SectionTitle>
-                        <Box>
-                          {cvData.certifications.items.map((cert) => (
-                            <EditableText key={cert.id} sx={{ position: 'relative' }}>
-                              <Typography variant="body2">{cert.content}</Typography>
-                              <EditControls className="edit-controls">
-                                <IconButton size="small" onClick={() => {
-                                  const newValue = prompt("Edit certification", cert.content);
-                                  if (newValue !== null) {
-                                    handleTextChange('certifications.items', cert.id, newValue);
-                                  }
-                                }}>
-                                  <Edit fontSize="small" />
-                                </IconButton>
-                                <IconButton size="small" onClick={() => handleDeleteItem('certifications.items', cert.id)}>
-                                  <Delete fontSize="small" />
-                                </IconButton>
-                              </EditControls>
-                            </EditableText>
-                          ))}
-                        </Box>
-                      </Section>
-                    </Grid>
-                  )}
-                </Grid>
-
-                {cvData.projects.length > 0 && (
-                  <Section>
+              <SectionCard>
+                <CardContent>
+                  <Box display="flex" justifyContent="space-between" alignItems="center">
                     <SectionTitle variant="h5" color="primary">
                       Projects
                     </SectionTitle>
-                    {cvData.projects.map((project) => (
-                      <Box key={project.id} mb={2} sx={{ position: 'relative' }}>
-                        <EditControls className="edit-controls" sx={{ top: -8, right: -8 }}>
-                          <IconButton size="small" onClick={() => handleDeleteItem('projects', project.id)}>
-                            <Delete fontSize="small" />
+                    <Tooltip title="Add project">
+                      <IconButton onClick={() => addNewItem('projects')}>
+                        <Add />
+                      </IconButton>
+                    </Tooltip>
+                  </Box>
+                  {cvData.projects.map((project) => (
+                    <Box key={project.id} mt={2} position="relative">
+                      <EditControls className="edit-controls" sx={{ top: -8, right: -8 }}>
+                        <IconButton size="small" onClick={() => {
+                          const newTitle = prompt("Edit project title", project.title);
+                          if (newTitle !== null) {
+                            handleObjectItemChange('projects', project.id, 'title', newTitle);
+                          }
+                        }}>
+                          <Edit fontSize="small" />
+                        </IconButton>
+                        <IconButton size="small" onClick={() => handleDeleteItem('projects', project.id)}>
+                          <Delete fontSize="small" />
+                        </IconButton>
+                      </EditControls>
+
+                      <EditableWrapper>
+                        <Typography fontWeight={600}>{project.title}</Typography>
+                      </EditableWrapper>
+
+                      <EditableWrapper>
+                        <Typography variant="body2" mt={1}>
+                          {project.description}
+                        </Typography>
+                        <EditControls className="edit-controls">
+                          <IconButton size="small" onClick={() => {
+                            const newValue = prompt("Edit description", project.description);
+                            if (newValue !== null) {
+                              handleObjectItemChange('projects', project.id, 'description', newValue);
+                            }
+                          }}>
+                            <Edit fontSize="small" />
                           </IconButton>
                         </EditControls>
-
-                        <EditableText>
-                          <Typography fontWeight={600}>{project.title}</Typography>
-                          <EditControls className="edit-controls">
-                            <IconButton size="small" onClick={() => {
-                              const newValue = prompt("Edit project title", project.title);
-                              if (newValue !== null) {
-                                handleObjectFieldChange('projects', project.id, 'title', newValue);
-                              }
-                            }}>
-                              <Edit fontSize="small" />
-                            </IconButton>
-                          </EditControls>
-                        </EditableText>
-
-                        <EditableText>
-                          <Typography variant="body2">{project.description}</Typography>
-                          <EditControls className="edit-controls">
-                            <IconButton size="small" onClick={() => {
-                              const newValue = prompt("Edit project description", project.description);
-                              if (newValue !== null) {
-                                handleObjectFieldChange('projects', project.id, 'description', newValue);
-                              }
-                            }}>
-                              <Edit fontSize="small" />
-                            </IconButton>
-                          </EditControls>
-                        </EditableText>
-                      </Box>
-                    ))}
-                  </Section>
-                )}
-              </Grid>
+                      </EditableWrapper>
+                    </Box>
+                  ))}
+                </CardContent>
+              </SectionCard>
             </Grid>
-          </CVContainer>
-        </Box>
+          </Grid>
+        </CVContainer>
       </Container>
     </ThemeProvider>
   );
