+++ src/utils/common.js
@@ -1,2 +1,2 @@
-// export const apiUrl = "http://localhost:9000/api/v1/portfolio";
-export const apiUrl = "http://192.168.1.5:9000/api/v1/portfolio";
+export const apiUrl = "http://localhost:9000/api/v1/portfolio";
+// export const apiUrl = "http://192.168.1.5:9000/api/v1/portfolio";
